refactor: rename components to PascalCase
